requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==42.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
psycopg2-binary == 2.9.9
prometheus-flask-exporter==0.23.0
python-json-logger==2.0.7